adds one item merchant analyst
